Combine some test classes to make things faster
